exploration of the dataset started.
loading the data and looking at its structure and different values.
conclusion: it seems like every datapoint has a dimension of 1916 and for each measurement we have a timestamp assigned. there are two main folders - one might be a training and the other one the test set (they have exactly the same struture). the training set covers approx 1year and the test set a couple of months.